% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\alias{decode.random_tree}
\alias{decode.encoder_forest}
\title{Decoder}
\usage{
decode(object, ...)

\method{decode}{random_tree}(object, terminal_ids, ...)

\method{decode}{encoder_forest}(object, terminal_ids, ...)
}
\arguments{
\item{object}{The object to use}

\item{...}{Additional arguments passed to individual methods.}

\item{terminal_ids}{Either a vector of terminal node ids (for an individual tree),
or a list of vectors of indvidual ids (produced by encoding from a while forest).}
}
\value{
A decoded object.
}
\description{
Calculate an decoding from an object (generic method)
}
