% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\alias{encode.random_tree}
\alias{encode.encoder_forest}
\title{Encoder}
\usage{
encode(object, ...)

\method{encode}{random_tree}(object, X, ...)

\method{encode}{encoder_forest}(object, X, ...)
}
\arguments{
\item{object}{The object to use}

\item{...}{Additional arguments passed to individual methods.}

\item{X}{The data to encode.}
}
\value{
An encoded object.
}
\description{
Calculate an encoding from an object (generic method)
}
