% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_forest.R
\name{encoder_forest}
\alias{encoder_forest}
\title{Encoder forest}
\usage{
encoder_forest(
  X,
  max_depth = 12,
  n_tree = 1000,
  subsample_size = max(126, nrow(X)/n_tree),
  resample = ifelse(subsample_size < (nrow(X)/n_tree), TRUE, FALSE),
  ...
)
}
\arguments{
\item{X}{The data to use.}

\item{max_depth}{The maximal depth of the tree.}

\item{n_tree}{The number of trees used.}

\item{subsample_size}{The number of rows allocated to each tree.}

\item{resample}{Whether to resample rows when creating samples of data for trees.}

\item{...}{Additional arguments (for future compatibility).}
}
\value{
A fitted encoder forest
}
\description{
Fit an encoder forest to your data
}
