% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_boostless_machine.R
\name{predict.greedy_boosting_machine}
\alias{predict.greedy_boosting_machine}
\title{Predict from terminal nodes of a tree in a greedy machine}
\usage{
\method{predict}{greedy_boosting_machine}(object, newdata, type = c("greedy", "all"), ...)
}
\arguments{
\item{object}{nada}

\item{newdata}{yada}

\item{type}{= c('greedy','all')}

\item{...}{prada}
}
\value{
results
}
\description{
Predict from terminal nodes of a tree in a greedy machine
}
