% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_uncertain_tree.R
\name{predict.random_tree}
\alias{predict.random_tree}
\alias{predict.encoder_forest}
\alias{predict.uncertain_tree}
\title{Predict from terminal nodes of a tree}
\usage{
\method{predict}{random_tree}(x, newdata, predict_fun = predictor_identity, ...)

\method{predict}{encoder_forest}(x, newdata, predict_fun = predictor_identity, ...)

\method{predict}{uncertain_tree}(x, newdata, predict_fun = predictor_identity, ...)
}
\arguments{
\item{x}{nada}

\item{newdata}{yada}

\item{predict_fun}{lada}

\item{...}{prada}
}
\value{
results
}
\description{
Predict from terminal nodes of a tree
}
